# services/patient-admin-service/docs/swagger.yaml
openapi: 3.0.0
info:
  title: ezClinic - Patient & Admin Service
  description: REST API for patient registration, profile management, medical reports, and admin operations. Owner - Minidu.
  version: "1.0.0"
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development Server
paths:
...
  /admins/register:
    post:
      summary: Register a new admin
      tags: [Admins]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminDto'
      responses:
        '201':
          description: Admin successfully registered
        '409':
          description: Email already exists
...
  /admins/{id}:
    get:
      summary: Get a specific admin
      tags: [Admins]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Admin found
    patch:
      summary: Update an admin
      tags: [Admins]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminDto'
      responses:
        '200':
          description: Admin updated
    delete:
      summary: Delete an admin
      tags: [Admins]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Admin deleted
...
  /admins/platform/patients/{id}:
    get:
      summary: Get a single patient with full details for Admin
      tags: [Admin Platform Operations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Complete patient record

    delete:
      summary: Delete a patient account (Admin operation)
      tags: [Admin Platform Operations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Patient account deleted
components:
  schemas:
...
    CreateAdminDto:
      type: object
      required:
        - name
        - email
      properties:
        auth0Id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
    UpdateAdminDto:
      type: object
      properties:
        name:
          type: string
...